+++ src/Messaging.Generic.Server/Options/GenericOptionsNames.cs
@@ -1,0 +1,12 @@
+﻿namespace Assistant.Net.Messaging.Options;
+
+/// <summary>
+///     Storage based server messaging handling options names.
+/// </summary>
+public static class GenericOptionsNames
+{
+    /// <summary>
+    ///     <see cref="MessagingClientOptions"/> instance name.
+    /// </summary>
+    public const string DefaultName = "server";
+}