+++ src/Messaging.Mongo.Server/GenericHandlingServerBuilderExtensions.cs
@@ -1,0 +1,52 @@
+﻿using Assistant.Net.Messaging.Options;
+using Assistant.Net.Options;
+using Assistant.Net.Storage;
+using Microsoft.Extensions.Configuration;
+using System;
+
+namespace Assistant.Net.Messaging;
+
+/// <summary>
+///     MongoDB remote message handling configuration extensions for a server.
+/// </summary>
+public static class GenericHandlingServerBuilderExtensions
+{
+    /// <summary>
+    ///     Configures server message handling to use MongoDB provider.
+    /// </summary>
+    /// <param name="builder"/>
+    /// <param name="configureOptions">The action used to configure the options.</param>
+    public static GenericHandlingServerBuilder UseMongo(this GenericHandlingServerBuilder builder, Action<MongoOptions> configureOptions)
+    {
+        builder.Services.ConfigureStorage(GenericOptionsNames.DefaultName, b => b
+            .UseMongo(configureOptions)
+            .UseMongoSingleProvider());
+        return builder;
+    }
+
+    /// <summary>
+    ///     Configures server message handling to use MongoDB provider.
+    /// </summary>
+    /// <param name="builder"/>
+    /// <param name="connectionString">The MongoDB connection string.</param>
+    public static GenericHandlingServerBuilder UseMongo(this GenericHandlingServerBuilder builder, string connectionString)
+    {
+        builder.Services.ConfigureStorage(GenericOptionsNames.DefaultName, b => b
+            .UseMongo(connectionString)
+            .UseMongoSingleProvider());
+        return builder;
+    }
+
+    /// <summary>
+    ///     Configures server message handling to use MongoDB provider.
+    /// </summary>
+    /// <param name="builder"/>
+    /// <param name="configuration">The application configuration values.</param>
+    public static GenericHandlingServerBuilder UseMongo(this GenericHandlingServerBuilder builder, IConfigurationSection configuration)
+    {
+        builder.Services.ConfigureStorage(GenericOptionsNames.DefaultName, b => b
+            .UseMongo(configuration)
+            .UseMongoSingleProvider());
+        return builder;
+    }
+}