+++ tests/Messaging.Sqlite.Tests/Fixtures/MessagingClientFixtureBuilder.cs
@@ -12,8 +12,8 @@ namespace Assistant.Net.Messaging.Sqlite.Tests.Fixtures;
 
 public class MessagingClientFixtureBuilder
 {
-    private readonly TestConfigureOptionsSource remoteSource = new();
     private readonly TestConfigureOptionsSource clientSource = new();
+    private readonly TestConfigureOptionsSource remoteSource = new();
 
     public MessagingClientFixtureBuilder()
     {
@@ -21,51 +21,47 @@ public class MessagingClientFixtureBuilder
             .AddMessagingClient(b => b
                 .RemoveInterceptor<CachingInterceptor>()
                 .RemoveInterceptor<RetryingInterceptor>()
-                .RemoveInterceptor<TimeoutInterceptor>())
-            .ConfigureSqliteHandlingClientOptions(o => o.ResponsePoll = new ConstantBackoff
+                .RemoveInterceptor<TimeoutInterceptor>()
+                .TimeoutIn(TimeSpan.FromSeconds(0.5)))
+            .ConfigureGenericHandlerProxyOptions(o => o.ResponsePoll = new ConstantBackoff
             {
                 Interval = TimeSpan.FromSeconds(0.05), MaxAttemptNumber = 5
             })
             .BindOptions(clientSource);
         RemoteHostBuilder = Host.CreateDefaultBuilder()
             .ConfigureServices(s => s
-                .AddSqliteMessageHandling(b => b
+                .AddGenericMessageHandling()
+                .ConfigureGenericMessageClient(o => o
                     .RemoveInterceptor<CachingInterceptor>()
                     .RemoveInterceptor<RetryingInterceptor>()
-                    .RemoveInterceptor<TimeoutInterceptor>())
-                .ConfigureSqliteHandlingServerOptions(o =>
+                    .RemoveInterceptor<TimeoutInterceptor>()
+                    .TimeoutIn(TimeSpan.FromSeconds(0.5)))
+                .ConfigureGenericHandlingServerOptions(o =>
                 {
                     o.InactivityDelayTime = TimeSpan.FromSeconds(0.005);
                     o.NextMessageDelayTime = TimeSpan.FromSeconds(0.001);
                 })
-                .BindOptions(SqliteOptionsNames.DefaultName, remoteSource));
+                .BindOptions(GenericOptionsNames.DefaultName, remoteSource));
     }
 
     public IServiceCollection Services { get; init; }
     public IHostBuilder RemoteHostBuilder { get; init; }
 
-    public MessagingClientFixtureBuilder UseSqlite(string connectionString)
+    public MessagingClientFixtureBuilder UseSqliteProvider(string connectionString)
     {
         Services.ConfigureMessagingClient(b => b
-            .UseSqlite(connectionString)
-            .UseSqliteProvider()
-            .TimeoutIn(TimeSpan.FromSeconds(0.5)));
-        RemoteHostBuilder.ConfigureServices(s => s
-            .ConfigureSqliteMessageHandling(b => b
-            .UseSqlite(connectionString)
-            .TimeoutIn(TimeSpan.FromSeconds(0.5))));
+            .UseSqlite(connectionString).UseSqliteProvider());
+        RemoteHostBuilder.ConfigureServices(s => s.ConfigureGenericMessageHandling(b => b
+            .UseSqlite(connectionString)));
         return this;
     }
 
     public MessagingClientFixtureBuilder UseSqliteSingleProvider(string connectionString)
     {
         Services.ConfigureMessagingClient(b => b
-            .UseSqlite(o => o.Connection(connectionString))
-            .UseSqliteSingleProvider()
-            .TimeoutIn(TimeSpan.FromSeconds(0.5)));
-        RemoteHostBuilder.ConfigureServices(s => s.ConfigureSqliteMessageHandling(b => b
-            .UseSqlite(o => o.Connection(connectionString))
-            .TimeoutIn(TimeSpan.FromSeconds(0.5))));
+            .UseSqlite(connectionString).UseSqliteSingleProvider());
+        RemoteHostBuilder.ConfigureServices(s => s.ConfigureGenericMessageHandling(b => b
+            .UseSqlite(connectionString)));
         return this;
     }
 
@@ -74,6 +70,7 @@ public class MessagingClientFixtureBuilder
         var messageType = typeof(THandler).GetMessageHandlerInterfaceTypes().FirstOrDefault()?.GetGenericArguments().First()
                           ?? throw new ArgumentException("Invalid message handler type.", nameof(THandler));
 
+        clientSource.Configurations.Add(o => o.AddGeneric(messageType));
         remoteSource.Configurations.Add(o =>
         {
             if (handler != null)
@@ -81,7 +78,6 @@ public class MessagingClientFixtureBuilder
             else
                 o.AddHandler(typeof(THandler));
         });
-        clientSource.Configurations.Add(o => o.AddSqlite(messageType));
         return this;
     }
 
@@ -90,6 +86,7 @@ public class MessagingClientFixtureBuilder
         var messageType = typeof(THandler).GetMessageHandlerInterfaceTypes().FirstOrDefault()?.GetGenericArguments().First()
                           ?? throw new ArgumentException("Invalid message handler type.", nameof(THandler));
 
+        clientSource.Configurations.Add(o => o.AddSingle(messageType));
         remoteSource.Configurations.Add(o =>
         {
             if (handler != null)
@@ -97,14 +94,13 @@ public class MessagingClientFixtureBuilder
             else
                 o.AddHandler(typeof(THandler));
         });
-        clientSource.Configurations.Add(o => o.Add(messageType));
         return this;
     }
 
     public MessagingClientFixtureBuilder AddMessageRegistrationOnly<TMessage>()
         where TMessage : IAbstractMessage
     {
-        clientSource.Configurations.Add(o => o.AddSqlite(typeof(TMessage)));
+        clientSource.Configurations.Add(o => o.AddGeneric(typeof(TMessage)));
         return this;
     }